Remove network interface
