+++ tests/flows/check_upgrade_access.go
@@ -5,21 +5,21 @@ import (
 	"math/big"
 
 	"github.com/ava-labs/subnet-evm/accounts/abi/bind"
-	"github.com/ava-labs/teleporter/tests/interfaces"
+	"github.com/ava-labs/teleporter/tests/network"
 	"github.com/ava-labs/teleporter/tests/utils"
 	"github.com/ethereum/go-ethereum/crypto"
 	. "github.com/onsi/gomega"
 )
 
-func CheckUpgradeAccess(network interfaces.Network) {
-	subnetInfo := network.GetPrimaryNetworkInfo()
-	fundedAddress, fundedKey := network.GetFundedAccountInfo()
+func CheckUpgradeAccess(n *network.LocalNetwork) {
+	subnetInfo := n.GetPrimaryNetworkInfo()
+	fundedAddress, fundedKey := n.GetFundedAccountInfo()
 
 	//
 	// Deploy TestMessenger to the subnet
 	//
 	ctx := context.Background()
-	teleporterAddress := network.GetTeleporterContractAddress()
+	teleporterAddress := n.GetTeleporterContractAddress()
 	_, testMessenger := utils.DeployTestMessenger(
 		ctx,
 		fundedKey,
