+++ tests/network/network.go
@@ -1,4 +1,4 @@
-package tests
+package network
 
 import (
 	"context"
@@ -36,8 +36,6 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 	. "github.com/onsi/gomega"
 )
-
-var _ interfaces.LocalNetwork = &LocalNetwork{}
 
 // Implements Network, pointing to the network setup in local_network_setup.go
 type LocalNetwork struct {
@@ -419,12 +417,6 @@ func (n *LocalNetwork) GetFundedAccountInfo() (common.Address, *ecdsa.PrivateKey
 
 func (n *LocalNetwork) IsExternalNetwork() bool {
 	return false
-}
-
-func (n *LocalNetwork) SupportsIndependentRelaying() bool {
-	// Messages can be relayed by the test application for local
-	// networks with connections to each node.
-	return true
 }
 
 func (n *LocalNetwork) RelayMessage(ctx context.Context,
