+++ tests/network/network_utils.go
@@ -1,0 +1,293 @@
+package network
+
+import (
+	"context"
+	"crypto/ecdsa"
+	"math/big"
+	"slices"
+	"time"
+
+	"github.com/ava-labs/avalanchego/ids"
+	"github.com/ava-labs/avalanchego/tests/fixture/tmpnet"
+	avalancheWarp "github.com/ava-labs/avalanchego/vms/platformvm/warp"
+	"github.com/ava-labs/subnet-evm/accounts/abi/bind"
+	"github.com/ava-labs/subnet-evm/core/types"
+	"github.com/ava-labs/subnet-evm/ethclient"
+	subnetEvmUtils "github.com/ava-labs/subnet-evm/tests/utils"
+	teleportermessenger "github.com/ava-labs/teleporter/abi-bindings/go/teleporter/TeleporterMessenger"
+	testmessenger "github.com/ava-labs/teleporter/abi-bindings/go/teleporter/tests/TestMessenger"
+	"github.com/ava-labs/teleporter/tests/interfaces"
+	"github.com/ava-labs/teleporter/tests/utils"
+	deploymentUtils "github.com/ava-labs/teleporter/utils/deployment-utils"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/log"
+
+	. "github.com/onsi/gomega"
+)
+
+// Issues txs to activate the proposer VM fork on the specified subnet index in the manager
+func setupProposerVM(ctx context.Context, fundedKey *ecdsa.PrivateKey, network *tmpnet.Network, subnetID ids.ID) {
+	subnetDetails := network.Subnets[slices.IndexFunc(
+		network.Subnets,
+		func(s *tmpnet.Subnet) bool { return s.SubnetID == subnetID },
+	)]
+
+	chainID := subnetDetails.Chains[0].ChainID
+
+	nodeURI, err := network.GetURIForNodeID(subnetDetails.ValidatorIDs[0])
+	Expect(err).Should(BeNil())
+	uri := utils.HttpToWebsocketURI(nodeURI, chainID.String())
+
+	client, err := ethclient.Dial(uri)
+	Expect(err).Should(BeNil())
+	chainIDInt, err := client.ChainID(ctx)
+	Expect(err).Should(BeNil())
+
+	err = subnetEvmUtils.IssueTxsToActivateProposerVMFork(ctx, chainIDInt, fundedKey, client)
+	Expect(err).Should(BeNil())
+}
+
+// Blocks until all validators specified in nodeURIs have reached the specified block height
+func waitForAllValidatorsToAcceptBlock(ctx context.Context, nodeURIs []string, blockchainID ids.ID, height uint64) {
+	cctx, cancel := context.WithTimeout(ctx, 10*time.Second)
+	defer cancel()
+	for i, uri := range nodeURIs {
+		chainAWSURI := utils.HttpToWebsocketURI(uri, blockchainID.String())
+		log.Debug("Creating ethclient for blockchain", "blockchainID", blockchainID.String(), "wsURI", chainAWSURI)
+		client, err := ethclient.Dial(chainAWSURI)
+		Expect(err).Should(BeNil())
+		defer client.Close()
+
+		// Loop until each node has advanced to >= the height of the block that emitted the warp log
+		for {
+			block, err := client.BlockByNumber(cctx, nil)
+			Expect(err).Should(BeNil())
+			if block.NumberU64() >= height {
+				log.Debug("Client accepted the block containing SendWarpMessage", "client", i, "height", block.NumberU64())
+				break
+			}
+		}
+	}
+}
+
+func ClearReceiptQueue(
+	ctx context.Context,
+	network *LocalNetwork,
+	fundedKey *ecdsa.PrivateKey,
+	source interfaces.SubnetTestInfo,
+	destination interfaces.SubnetTestInfo,
+) {
+	Expect(network.IsExternalNetwork()).Should(BeFalse())
+	outstandReceiptCount := utils.GetOutstandingReceiptCount(source, destination.BlockchainID)
+	for outstandReceiptCount.Cmp(big.NewInt(0)) != 0 {
+		log.Info("Emptying receipt queue", "remainingReceipts", outstandReceiptCount.String())
+		// Send message from Subnet B to Subnet A to trigger the "regular" method of delivering receipts.
+		// The next message from B->A will contain the same receipts that were manually sent in the above steps,
+		// but they should not be processed again on Subnet A.
+		sendCrossChainMessageInput := teleportermessenger.TeleporterMessageInput{
+			DestinationBlockchainID: destination.BlockchainID,
+			DestinationAddress:      common.HexToAddress("0x1111111111111111111111111111111111111111"),
+			RequiredGasLimit:        big.NewInt(1),
+			FeeInfo: teleportermessenger.TeleporterFeeInfo{
+				FeeTokenAddress: common.Address{},
+				Amount:          big.NewInt(0),
+			},
+			AllowedRelayerAddresses: []common.Address{},
+			Message:                 []byte{1, 2, 3, 4},
+		}
+
+		// This message will also have the same receipts as the previous message
+		receipt, _ := utils.SendCrossChainMessageAndWaitForAcceptance(
+			ctx, source, destination, sendCrossChainMessageInput, fundedKey)
+
+		// Relay message
+		network.RelayMessage(ctx, receipt, source, destination, true)
+
+		outstandReceiptCount = utils.GetOutstandingReceiptCount(source, destination.BlockchainID)
+	}
+	log.Info("Receipt queue emptied")
+}
+
+// Returns Receipt for the transaction unlike TeleporterRegistry version since this is a non-teleporter case
+// and we don't want to add the ValidatorSetSig ABI to the subnetInfo
+func ExecuteValidatorSetSigCallAndVerify(
+	ctx context.Context,
+	network *LocalNetwork,
+	source interfaces.SubnetTestInfo,
+	destination interfaces.SubnetTestInfo,
+	validatorSetSigAddress common.Address,
+	senderKey *ecdsa.PrivateKey,
+	unsignedMessage *avalancheWarp.UnsignedMessage,
+	expectSuccess bool,
+) *types.Receipt {
+	signedWarpMsg := network.GetSignedMessage(ctx, source, destination, unsignedMessage.ID())
+	log.Info("Got signed warp message", "messageID", signedWarpMsg.ID())
+
+	signedPredicateTx := utils.CreateExecuteCallPredicateTransaction(
+		ctx,
+		signedWarpMsg,
+		validatorSetSigAddress,
+		senderKey,
+		destination,
+	)
+
+	// Wait for tx to be accepted and verify events emitted
+	if expectSuccess {
+		return utils.SendTransactionAndWaitForSuccess(ctx, destination, signedPredicateTx)
+	}
+	return utils.SendTransactionAndWaitForFailure(ctx, destination, signedPredicateTx)
+}
+
+func AddProtocolVersionAndWaitForAcceptance(
+	ctx context.Context,
+	network *LocalNetwork,
+	subnet interfaces.SubnetTestInfo,
+	newTeleporterAddress common.Address,
+	senderKey *ecdsa.PrivateKey,
+	unsignedMessage *avalancheWarp.UnsignedMessage,
+) {
+	signedWarpMsg := network.GetSignedMessage(ctx, subnet, subnet, unsignedMessage.ID())
+	log.Info("Got signed warp message", "messageID", signedWarpMsg.ID())
+
+	// Construct tx to add protocol version and send to destination chain
+	signedTx := utils.CreateAddProtocolVersionTransaction(
+		ctx,
+		signedWarpMsg,
+		subnet.TeleporterRegistryAddress,
+		senderKey,
+		subnet,
+	)
+
+	curLatestVersion, err := subnet.TeleporterRegistry.LatestVersion(&bind.CallOpts{})
+	Expect(err).Should(BeNil())
+	expectedLatestVersion := big.NewInt(curLatestVersion.Int64() + 1)
+
+	// Wait for tx to be accepted, and verify events emitted
+	receipt := utils.SendTransactionAndWaitForSuccess(ctx, subnet, signedTx)
+	addProtocolVersionEvent, err := utils.GetEventFromLogs(receipt.Logs, subnet.TeleporterRegistry.ParseAddProtocolVersion)
+	Expect(err).Should(BeNil())
+	Expect(addProtocolVersionEvent.Version.Cmp(expectedLatestVersion)).Should(Equal(0))
+	Expect(addProtocolVersionEvent.ProtocolAddress).Should(Equal(newTeleporterAddress))
+
+	versionUpdatedEvent, err := utils.GetEventFromLogs(receipt.Logs, subnet.TeleporterRegistry.ParseLatestVersionUpdated)
+	Expect(err).Should(BeNil())
+	Expect(versionUpdatedEvent.OldVersion.Cmp(curLatestVersion)).Should(Equal(0))
+	Expect(versionUpdatedEvent.NewVersion.Cmp(expectedLatestVersion)).Should(Equal(0))
+}
+
+func (network *LocalNetwork) GetTwoSubnets() (
+	interfaces.SubnetTestInfo,
+	interfaces.SubnetTestInfo,
+) {
+	subnets := network.GetSubnetsInfo()
+	Expect(len(subnets)).Should(BeNumerically(">=", 2))
+	return subnets[0], subnets[1]
+}
+
+func SendExampleCrossChainMessageAndVerify(
+	ctx context.Context,
+	network *LocalNetwork,
+	source interfaces.SubnetTestInfo,
+	sourceExampleMessenger *testmessenger.TestMessenger,
+	destination interfaces.SubnetTestInfo,
+	destExampleMessengerAddress common.Address,
+	destExampleMessenger *testmessenger.TestMessenger,
+	senderKey *ecdsa.PrivateKey,
+	message string,
+	expectSuccess bool,
+) {
+	// Call the example messenger contract on Subnet A
+	optsA, err := bind.NewKeyedTransactorWithChainID(senderKey, source.EVMChainID)
+	Expect(err).Should(BeNil())
+	tx, err := sourceExampleMessenger.SendMessage(
+		optsA,
+		destination.BlockchainID,
+		destExampleMessengerAddress,
+		common.BigToAddress(common.Big0),
+		big.NewInt(0),
+		testmessenger.SendMessageRequiredGas,
+		message,
+	)
+	Expect(err).Should(BeNil())
+
+	// Wait for the transaction to be mined
+	receipt := utils.WaitForTransactionSuccess(ctx, source, tx.Hash())
+
+	event, err := utils.GetEventFromLogs(receipt.Logs, source.TeleporterMessenger.ParseSendCrossChainMessage)
+	Expect(err).Should(BeNil())
+	Expect(event.DestinationBlockchainID[:]).Should(Equal(destination.BlockchainID[:]))
+
+	teleporterMessageID := event.MessageID
+
+	//
+	// Relay the message to the destination
+	//
+	receipt = network.RelayMessage(ctx, receipt, source, destination, true)
+
+	//
+	// Check Teleporter message received on the destination
+	//
+	delivered, err := destination.TeleporterMessenger.MessageReceived(
+		&bind.CallOpts{}, teleporterMessageID,
+	)
+	Expect(err).Should(BeNil())
+	Expect(delivered).Should(BeTrue())
+
+	if expectSuccess {
+		// Check that message execution was successful
+		messageExecutedEvent, err := utils.GetEventFromLogs(
+			receipt.Logs,
+			destination.TeleporterMessenger.ParseMessageExecuted,
+		)
+		Expect(err).Should(BeNil())
+		Expect(messageExecutedEvent.MessageID[:]).Should(Equal(teleporterMessageID[:]))
+	} else {
+		// Check that message execution failed
+		messageExecutionFailedEvent, err := utils.GetEventFromLogs(
+			receipt.Logs,
+			destination.TeleporterMessenger.ParseMessageExecutionFailed,
+		)
+		Expect(err).Should(BeNil())
+		Expect(messageExecutionFailedEvent.MessageID[:]).Should(Equal(teleporterMessageID[:]))
+	}
+
+	//
+	// Verify we received the expected string
+	//
+	_, currMessage, err := destExampleMessenger.GetCurrentMessage(&bind.CallOpts{}, source.BlockchainID)
+	Expect(err).Should(BeNil())
+	if expectSuccess {
+		Expect(currMessage).Should(Equal(message))
+	} else {
+		Expect(currMessage).ShouldNot(Equal(message))
+	}
+}
+
+// Deploys a new version of Teleporter and returns its address
+// Does NOT modify the global Teleporter contract address to provide greater testing flexibility.
+func DeployNewTeleporterVersion(
+	ctx context.Context,
+	network *LocalNetwork,
+	fundedKey *ecdsa.PrivateKey,
+	teleporterByteCodeFile string,
+) common.Address {
+	contractCreationGasPrice := (&big.Int{}).Add(deploymentUtils.GetDefaultContractCreationGasPrice(), big.NewInt(1))
+	teleporterDeployerTransaction,
+		_,
+		teleporterDeployerAddress,
+		teleporterContractAddress,
+		err := deploymentUtils.ConstructKeylessTransaction(
+		teleporterByteCodeFile,
+		false,
+		contractCreationGasPrice,
+	)
+	Expect(err).Should(BeNil())
+
+	network.DeployTeleporterContractToAllChains(
+		teleporterDeployerTransaction,
+		teleporterDeployerAddress,
+		teleporterContractAddress,
+		fundedKey,
+	)
+	return teleporterContractAddress
+}
